Fix issue when invoking from project subfolder.
When the lein oneoff command was invoked from a subfolder of a parent directory
that contained a project.clj file, the path to the oneoff script would not be correct
because leiningen changes the current working directory to the root of the project.

Use "leiningen.original.pwd" property to fix that issue.